WA-232 Test localStorage's tokens
